done posts version1

+++ hw9/src/edu/cmu/cs214/hw9/db/Post.java
@@ -4,9 +4,9 @@ public class Post {
 	private String content;
 	private String email;
 	private int is_status;
-	private int date_added;
+	private long date_added;
 	
-	public Post(String email, String content, int is_status, int date_added){
+	public Post(String email, String content, int is_status, long date_added){
 		this.email = email;
 		this.content = content;
 		this.is_status = is_status;
@@ -25,7 +25,7 @@ public class Post {
 		return is_status;
 	}
 	
-	public int getDateAdded(){
+	public long getDateAdded(){
 		return date_added;
 	}
 }
